@@ -29,21 +29,21 @@
     "@polkadot/hw-ledger": "^9.0.1",
     "@polkadot/keyring": "^9.0.1",
     "@polkadot/networks": "^9.0.1",
-    "@polkadot/react-identicon": "^2.0.1",
-    "@polkadot/react-qr": "^2.0.1",
-    "@polkadot/types": "^8.1.1",
-    "@polkadot/types-augment": "^8.1.1",
-    "@polkadot/ui-keyring": "^2.0.1",
-    "@polkadot/ui-settings": "^2.0.1",
+    "@polkadot/react-identicon": "^2.0.2-0",
+    "@polkadot/react-qr": "^2.0.2-0",
+    "@polkadot/types": "^8.2.2-2",
+    "@polkadot/types-augment": "^8.2.2-2",
+    "@polkadot/ui-keyring": "^2.0.2-0",
+    "@polkadot/ui-settings": "^2.0.2-0",
     "@polkadot/util": "^9.0.1",
     "@polkadot/util-crypto": "^9.0.1",
     "file-saver": "^2.0.5",
     "i18next": "^21.6.16",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-dropzone": "^11.7.1",
     "react-i18next": "^11.16.7",
-    "react-is": "^18.0.0",
+    "react-is": "^18.1.0",
     "react-router": "^5.3.1",
     "react-router-dom": "^5.3.1",
     "styled-components": "^5.3.5"
