@@ -4,7 +4,7 @@
 import './i18n/i18n';
 
 import React, { Suspense } from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { HashRouter } from 'react-router-dom';
 
 import { View } from './components';
@@ -16,14 +16,13 @@ export default function createView (Entry: React.ComponentType, rootId = 'root')
     throw new Error(`Unable to find element with id '${rootId}'`);
   }
 
-  ReactDOM.render(
+  createRoot(rootElement).render(
     <Suspense fallback='...'>
       <View>
         <HashRouter>
           <Entry />
         </HashRouter>
       </View>
-    </Suspense>,
-    rootElement
+    </Suspense>
   );
 }
